+++ 1w-tdd/src/main/java/io/hhplus/tdd/point/vo/TransactionType.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point.dto;
+package io.hhplus.tdd.point.vo;
 
 
 /**
