+++ 1w-tdd/src/main/java/io/hhplus/tdd/point/vo/UserPoint.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd.point.dto;
+package io.hhplus.tdd.point.vo;
 
 public record UserPoint(
         long id,
