+++ 1w-tdd/src/main/java/io/hhplus/tdd/util/ApiControllerAdvice.java
@@ -1,4 +1,4 @@
-package io.hhplus.tdd;
+package io.hhplus.tdd.util;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
